Expose GLFW instance extension provider
Use it when creating instances in GLFW examples, and make Vulkan
surface creation accept const instances to avoid const_casts.
