+++ integrations/gfx_glfw/CMakeLists.txt
@@ -18,10 +18,16 @@ else()
     target_sources(GraphicsGLFW PRIVATE "src/gfx_glfw.cpp")
 endif()
 
+if (GFX_BUILD_VULKAN)
+    target_sources(GraphicsGLFW PRIVATE "src/GlfwInstanceExtension.cpp")
+endif()
+
 if (GFX_BUILD_METAL)
+    target_compile_definitions(GraphicsGLFW PRIVATE "GFX_BUILD_METAL")
     target_compile_options(GraphicsGLFW PRIVATE "-fobjc-arc")
 endif()
 if (GFX_BUILD_VULKAN)
+    target_compile_definitions(GraphicsGLFW PRIVATE "GFX_BUILD_VULKAN")
     target_compile_definitions(GraphicsGLFW PRIVATE "VULKAN_HPP_DISPATCH_LOADER_DYNAMIC=1")
     target_compile_definitions(GraphicsGLFW PRIVATE "VULKAN_HPP_NO_CONSTRUCTORS=1")
     target_compile_definitions(GraphicsGLFW PRIVATE "VULKAN_HPP_HANDLES_MOVE_EXCHANGE=1")
