+++ integrations/gfx_imgui/CMakeLists.txt
@@ -25,6 +25,10 @@ if (GFX_BUILD_VULKAN)
     target_sources(GraphicsImGui PRIVATE "src/imgui_impl_vulkan.h" "src/imgui_impl_vulkan.cpp")
 endif()
 
+if (GFX_BUILD_METAL)
+    target_compile_definitions(GraphicsImGui PRIVATE "GFX_BUILD_METAL")
+    target_compile_options(GraphicsImGui PRIVATE "-fobjc-arc")
+endif()
 if (GFX_BUILD_VULKAN)
     target_compile_definitions(GraphicsImGui PRIVATE "GFX_BUILD_VULKAN")
     target_compile_definitions(GraphicsImGui PRIVATE "IMGUI_IMPL_VULKAN_NO_PROTOTYPES")
@@ -32,11 +36,6 @@ if (GFX_BUILD_VULKAN)
     target_compile_definitions(GraphicsImGui PRIVATE "VULKAN_HPP_NO_CONSTRUCTORS=1")
     target_compile_definitions(GraphicsImGui PRIVATE "VULKAN_HPP_HANDLES_MOVE_EXCHANGE=1")
     set_source_files_properties("${CMAKE_CURRENT_LIST_DIR}/src/imgui_impl_vulkan.cpp" PROPERTIES SKIP_PRECOMPILE_HEADERS ON)
-endif()
-
-if (GFX_BUILD_METAL)
-    target_compile_definitions(GraphicsImGui PRIVATE "GFX_BUILD_METAL")
-    target_compile_options(GraphicsImGui PRIVATE "-fobjc-arc")
 endif()
 
 if (GFX_BUILD_VULKAN)
